Factor PanicMessage out of rpc code in proc_macro
